// lib/types/json.d.ts
export interface JSONCommandInteraction extends JSONInteraction {
    appPermissions?: JSONPermission;
    channelID: string;
    data: ApplicationCommandInteractionData;
    guildID?: string;
    guildLocale?: string;
    locale: string;
    member?: JSONMember;
    type: InteractionTypes.APPLICATION_COMMAND;
    user: JSONUser;
}

export interface JSONChatInputCommandInteraction extends JSONCommandInteraction {
    options?: InteractionOptionsWrapper;
}
// ...
export interface JSONInteraction extends JSONBase {
    applicationID: string;
    token: string;
    type: InteractionTypes;
    version: 1;
}
// ...
export interface JSONMember extends JSONBase {
    avatar: string | null;
    communicationDisabledUntil: number | null;
    deaf: boolean;
    flags?: number;
    guildID: string;
    isPending?: boolean;
    joinedAt: number | null;
    mute: boolean;
    nick: string | null;
    pending: boolean;
    premiumSince: number | null;
    presence?: Presence;
    roles: Array<string>;
    user: JSONUser;
}
// ...
export interface JSONPermission {
    allow: string;
    deny: string;
}
// ...
export interface JSONUser extends JSONBase {
    accentColor?: number | null;
    avatar: string | null;
    banner?: string | null;
    bot: boolean;
    discriminator: string;
    publicFlags: number;
    system: boolean;
    username: string;
}